Move to QuTiP-based calculations
This makes everything now return qutip.Qobj types with the tensor
product explicitly represented.  This should hopefully remove any
possible confusion with what is represented, and allows for easier
mathematical use of the objects without me having to recreate the wheel
constantly.

For example, there is a method Qobj.ptrace() which is perfect for
calculating the effects of measurements in the ion trap.