Allow Sequences to take (times, phases) pairs
The Sequence.op and Sequence.d_op methods can now take either a single
interleaved list of [time0, phase0, time1, ...], or a pair of lists with
the times and phases separated, i.e.
    [time0, time1, ...], [phase0, phase1, ...]
This should be a bit more user-friendly, and a bit easier to read when
dealing with them in an interactive session.