Remove Laser class
The Laser class was very misleadingly named, and made it a nuisance to
implement the ability to change the base Rabi frequency, the detuning or
the Lamb--Dicke parameter of a sideband that already existed.

In removing the Laser class, we now make the Sideband class (and by
extension the Sequence class) able to be updated with new detunings and
other parameters without reallocating the class.  Aside from the reduced
memory requirements, this also represents an approximately 12% speedup
over the naive complete reallocation.